feat : The program gives out a user by id.

+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryUserStorage.java
@@ -27,7 +27,7 @@ public class InMemoryUserStorage implements UserStorage {
     }
 
     @Override
-    public User findUserById(Long userId) {
+    public User findUserById(Integer userId) {
         return users.get(userId);
     }
